Improve visibility of unexpected pixel values
This moves the DiffShower into its own class.
When the PixelComparator finds mismatched pixels, meaning that the provided
value differs from the expected value, it will open the DiffShower GUI to
show the mismatched pixels. This is much easier to debug than a generic
"it's broken."

The DiffShower has four rows.
The first row shows the indices of the pixels.
The second row shows the expected pixel values.
The third row shows the actual pixel values.
The fourth row shows an overview of what is different. Non-differing pixels
will have a white square in the fourth row. Differing pixels will have a
red square.
