+++ Robot/neopixel-leds/src/test/java/org/usfirst/frc/team4999/tools/PixelComparator.java
@@ -10,81 +10,13 @@ import java.nio.file.Path;
 import java.util.Arrays;
 import java.util.Vector;
 
-import javax.swing.JComponent;
-import javax.swing.JFrame;
+import org.usfirst.frc.team4999.tools.gui.DiffShower;
 
 import java.awt.Color;
-import java.awt.Dimension;
-import java.awt.Graphics;
-import java.awt.Graphics2D;
-import java.awt.Rectangle;
-import java.awt.event.WindowAdapter;
-import java.awt.event.WindowEvent;
 
 import static org.junit.Assert.fail;
 
 public class PixelComparator implements BufferDisplay.BufferUpdateListener {
-
-    private static class DifferenceShower {
-
-        private static Object lock = new Object();
-
-        public static void showDifference(Color[] expected, Color[] actual) {
-            final int pixel_size = 20;
-            final int txt_offset = 4;
-            JComponent component = new JComponent() {
-                private static final long serialVersionUID = 2L;
-                @Override
-                public void paintComponent(Graphics gd) {
-                    Graphics2D g = (Graphics2D) gd;
-                    g.setPaint(Color.BLACK);
-                    for(int i = 0; i < expected.length; i++) {
-                        Rectangle rect = new Rectangle(i * pixel_size, 0, pixel_size, pixel_size);
-                        String num = Integer.toString(i);
-                        g.draw(rect);
-                        g.drawString(num, i*pixel_size + txt_offset, pixel_size - txt_offset);
-                    }
-                    for(int i = 0; i < expected.length; i++) {
-                        Rectangle rect = new Rectangle(i * pixel_size, pixel_size, pixel_size, pixel_size);
-                        g.setPaint(expected[i]);
-                        g.fill(rect);
-                    }
-                    for(int i = 0; i < actual.length; i++) {
-                        Rectangle rect = new Rectangle(i * pixel_size, pixel_size * 2, pixel_size, pixel_size);
-                        g.setPaint(actual[i]);
-                        g.fill(rect);
-                    }
-                }
-            };
-            component.setPreferredSize(new Dimension(Math.max(expected.length, actual.length) * pixel_size, pixel_size * 3));
-
-            JFrame frame = new JFrame();
-            frame.add(component);
-            frame.setDefaultCloseOperation(JFrame.HIDE_ON_CLOSE);
-            frame.pack();
-            frame.addWindowListener(new WindowAdapter() {
-                @Override
-                public void windowClosing(WindowEvent arg) {
-                    synchronized(lock) {
-                        frame.setVisible(false);
-                        lock.notifyAll();
-                    }
-                }
-            });
-
-            frame.setVisible(true);
-
-            synchronized(lock) {
-                try {
-                    lock.wait();
-                } catch (InterruptedException e) {
-                    e.printStackTrace();
-                }
-            }
-
-            frame.dispose();
-        }
-    }
 
     private Vector<Color[]> displayHistory;
 
@@ -188,13 +120,13 @@ public class PixelComparator implements BufferDisplay.BufferUpdateListener {
 
             if(curr.length != currKey.length) {
                 fail(String.format("Buffer sizes differ: expected %d, actual %d\n", currKey, curr));
-                if(!headless) DifferenceShower.showDifference(currKey, curr);
+                if(!headless) DiffShower.showDifference(currKey, curr);
                 return;
             }
 
             for(int j = 0; j < curr.length; j++) {
                 if(!curr[j].equals(currKey[j])) {
-                    if(!headless) DifferenceShower.showDifference(currKey, curr);
+                    if(!headless) DiffShower.showDifference(currKey, curr);
                     fail(String.format("Buffer pixels differ: (frame:%d color:%d) %s != %s\n", i, j, curr[j], currKey[j]));
                     return;
                 }
